refactor(account): Move accout to benches, Use PassEntry for password management
refactor: Removed unused imports and code

refactor(modules): made pass/utils for utility funtions
