+++ src/store/password_entry.rs
@@ -1,6 +1,5 @@
-use std::{num::NonZeroU32, slice};
+use std::num::NonZeroU32;
 
-use super::account::Account;
 use ring::{
     pbkdf2,
     rand::{SecureRandom, SystemRandom},
@@ -8,8 +7,10 @@ use ring::{
 use serde::{Deserialize, Serialize};
 use serde_encrypt::{
     serialize::impls::BincodeSerializer, shared_key::SharedKey, traits::SerdeEncryptSharedKey,
-    AsSharedKey, EncryptedMessage,
+    EncryptedMessage,
 };
+
+use crate::pass::util::{derive_encryption_key, generate_random_password};
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq, PartialOrd)]
 struct Password {
@@ -32,11 +33,11 @@ struct PasswordEntry {
     /// Name of service/email/website for which password is stored
     service: String,
 
+    /// Username or Identifer used for that service
+    username: String,
+
     /// Password for service
     password: Password,
-
-    /// Additional details like username, notes etc
-    details: Option<Account>,
 }
 
 impl SerdeEncryptSharedKey for PasswordEntry {
@@ -45,17 +46,17 @@ impl SerdeEncryptSharedKey for PasswordEntry {
 
 impl PasswordEntry {
     /// Function for initialising entry of a password by taking details of it
-    pub fn new(service: String, password: Option<Vec<u8>>, details: Option<Account>) -> Self {
+    pub fn new(service: String, username: String, password: Option<Vec<u8>>) -> Self {
         Self {
             service,
+            username,
             password: Password::new(password),
-            details,
         }
     }
 
     /// Encrypt the entry
     pub fn encrypt_entry(&self, master_pass: &Vec<u8>) -> Result<Vec<u8>, serde_encrypt::Error> {
-        let key = derive_encrytion_key(master_pass, "Salt".as_bytes());
+        let key = derive_encryption_key(master_pass, "Salt".as_bytes());
         let key = SharedKey::new(key);
 
         let encrypted_content = self.encrypt(&key)?;
@@ -67,71 +68,13 @@ impl PasswordEntry {
         content: &Vec<u8>,
         master_pass: &Vec<u8>,
     ) -> Result<Self, serde_encrypt::Error> {
-        let key = derive_encrytion_key(master_pass, "Salt".as_bytes());
+        let key = derive_encryption_key(master_pass, "Salt".as_bytes());
         let key = SharedKey::new(key);
 
         let encrypted_content = EncryptedMessage::deserialize(content.to_owned())?;
         let decrypted_content = PasswordEntry::decrypt_owned(&encrypted_content, &key)?;
         Ok(decrypted_content)
     }
-}
-
-// Derive a encryption key from master password & salt
-pub fn derive_encrytion_key(master_pass: &Vec<u8>, salt: &[u8]) -> [u8; 32] {
-    let mut encryption_key = [0_u8; 32];
-
-    pbkdf2::derive(
-        pbkdf2::PBKDF2_HMAC_SHA256,
-        NonZeroU32::new(600_000).unwrap(),
-        salt,
-        master_pass,
-        &mut encryption_key,
-    );
-
-    encryption_key
-}
-
-// Genrerate a random salt using Rng
-pub fn get_random_salt() -> [u8; 16] {
-    let mut salt = [0u8; 16];
-    let r = SystemRandom::new();
-    r.fill(&mut salt).unwrap();
-    salt
-}
-
-// Function to verify the master password is strong enough
-pub fn is_strong_password(password: &str) -> bool {
-    // Check if the password length is at least 8 characters
-    if password.len() < 8 {
-        return false;
-    }
-
-    let has_lowercase = password.chars().any(|c| c.is_ascii_lowercase());
-    let has_uppercase = password.chars().any(|c| c.is_ascii_uppercase());
-    let has_digit = password.chars().any(|c| c.is_ascii_digit());
-    let has_special = password
-        .chars()
-        .any(|c| !c.is_alphanumeric() && !c.is_whitespace());
-
-    return has_lowercase && has_uppercase && has_digit && has_special;
-}
-
-pub fn generate_random_password(length: u8) -> String {
-    use rand::Rng;
-    const CHARSET: &[u8] = b"ABCDEFGHIJKLMNOPQRSTUVWXYZ\
-                        abcdefghijklmnopqrstuvwxyz\
-                        0123456789)(*&^%$#@!~";
-    let password_len: u8 = length;
-    let mut rng = rand::thread_rng();
-
-    let password: String = (0..password_len)
-        .map(|_| {
-            let idx = rng.gen_range(0..CHARSET.len());
-            CHARSET[idx] as char
-        })
-        .collect();
-
-    password
 }
 
 #[cfg(test)]
@@ -151,8 +94,8 @@ mod test {
     fn encrypt_decrypt_works() -> Result<(), serde_encrypt::Error> {
         let any_entry = PasswordEntry::new(
             String::from("Netflix"),
+            String::new(),
             Some(b"GxM4B7PDBe3NVY!Yj7A&&hvPs!ssJ3^q".to_vec()),
-            None,
         );
         dbg!(&any_entry);
